This fixes issue 29.
Stage 2 on the reconstruct the active formatting elements pass was ambiguous. Instead of checking for the *exact same element* on the stack of open nodes (i.e. search for the pointer) I should have been looking for a similar element - i.e. same tag name, but not the same element.

Also removes an unnecessary if (inScope) that is just before a while (inScope).