Update Image so that it works with DS
The DS version of libpng seemed to align the 3-byte colour structure.
Catch this case and make use of it with the channels() function in Image.

Update the test case and ImageComponent to use this new API.
Make the test.png a colormap png, since this triggered the problem.

Update the images SConscript to copy the images pc binary to
the test directory, gitignore it.