Show visited links in a different colour.
Update the Cache so that it stores file names in memory and doesn't have
to check the disk each time. Remove duplicate clean() code.

Add a URI method to get the raw integer CRC of a URI.

Change ViewRender to tell the textArea if the link that is added has been
visited already or not.

Finally, in RichTextArea, set the link colour for visited links to purple.