Fix holding down on the scroll bar.
Stylus class changes broke the scroll bar. Now when the user holds down
the stylus, the scroll bar will keep scrolling instead of stopping.

Also allowed the user to fall outside the bounds of the scroller once she
has hit inside it - this is how most other scrollbars seem to work and is
less irritating.