Add basic bookmarking features.
Refactor the Toolbar code to allow toolbar states. The regular icons
are now the "browsing toolbar". Add a bookmark icon to the main toolbar.
When in bookmark mode, show the "bookmark toolbar".

The bookmark toolbar has a back button and an add button.
Clicking add writes out the URL/Title of the last visited page. Clicking
back returns to that page. Showing BMs doesn't change the history,
despite the bookmarks page being a crude HTML page.

Other changes to support this are to UnicodeString, which now encodes UTF8
to char *. This means it is possible to add UTF8 encoded bookmarks. Add
a manual test case for this.