Join the online status and spinner into one icon, add prefs icon, refactor.
The wifi connection icon and the spinner are pretty much mutually
exclusive, there's no need to see them both together.

Added the preferences icon for edditing "stuff" - nothing done yet.

The steps to set a new toolbar are to hide the old one, set the new value
and show the new toolbar. This has been factored out to a new method. Also,
fix a bug in Toolbar that had cropped up from the last commit.