Improve save-as to check mime type via HEAD request.
This implements the basics for better downloading. Each page is asked for
its HEAD-ers first, then based on this the file is either shown, or
downloaded. The user is asked for a file name when downloading, defaulting
to the last part of the URL.

The HttpClient API is changed to accept just URI ref.  Instead of passing
in server, port and URI, just pass in the URI to the HttpClient
constructor.

When a save file is cancelled, it calls Document::setStatus(LOADED).  This
causes the "content not viewable" screen to be shown when the download ends
or if cancel is pressed.

A way to keep track of the amount of binary data downloaded is also added,
currently no progress meter is there but the back end stuff is in place.