Default download directory configuration.
Add a way to extract configuration values and show them in the config
screen. Use %cfg:KEY% and KEY is replaced by the config value. Add a
test case for this.

Update the configuration html with the download option. Update config.ini
with missing keys, if the key is new.