Add a way to patch files using DLDI.
On the NDS, copy the current DLDI patch from memory into the file in
question. This uses the code from Chishm's NDS Loader to get the DLDI patch
address.

On the PC, use a dldi file from $DLDIPATH as the patch.

The actual patching uses the DLDI patcher tool from DevkitPro.

Add a test file for patching.