Improve updater.
This changes the updater into a real "view". It keeps track of the state,
shows crappy messages about what is going on and allows the user to bail
out of downloading the upgrade.

version.c now holds a real version number instead of a date string.

RichTexArea fix for new line after buttons. This may break things.

Translate a bit of the Keyboard. Add more values to the en.txt file.