Add autoconf-like configure scripts.
These configure scripts check for "scons" and give advice on how to
install it. They then generate Makefiles in all subdirs that have
SConscripts, so "make" works in any directory. This hopefully gives the
impression of an easier build system...

Fix grit bug when DEVKITARM set, but grit ain't found.