Refactor popup menus.
Refactoring of the EditPopup and LinkHandler code.
Change EditPopup to use new Popupmenu class.
Removes the duplicate code making LinkHandler a PopupMenu.

After adding items to the menu, the layout has to be redone. Add this
method that lets sub classes fiddle the layout correctly without having
to #include the world.