Fix dash characters shown as malformed glyphs.
Using isblank() caused some wide characters to be marked as whitespace,
for example the ndash or mdash (0x2013, 0x2014) character.

Add these dashes to the font, plus some others that are used quite often.
Add a hack for ISO-8859-1, which uses some code page stuff. Some websites
use &#151; instead of &ndash; which caused the glyph to be shown munged.

Updated the render examples for this case as well as adding a new html
with all entities on it.