Fix top screen garbage after DLDI patch.
After patching with DLDI, the top screen was left with the memory dump
from the patch routine. Touching the lower screen triggered a redraw, but
the effect was so nasty it'd scare people.

Now the progress bar can be set invisible, so use that mechanism to trigger
a redraw on the top screen.