Fix for scrolling way off the bottom of the page.
Pressing right to scroll down often scrolled miles past the end of the
page. This was due to assumptions in the ScrollPane that the main text
was the first child. Change this so that it is the end of the last child
element.

Also update the long-form example so that it reproduces the bug.