Fix black lines at the end of some images.
Due to the way Images are scaled (using 24:8 fixed point) there is a good
chance of rounding error when scaling vertically. Luckily the
m_currentLine variable holds the actual real drawn height of the image.
Store this as the height value.

Make sure m_currentLine is initialised to 0 in Image.cpp

Update unit test as the image is slightly smaller than advertised.