waf: Move build to the top level directory
Instead of configuring each directory separately, now you run configure
and build from the top level. This means dependencies between libndspp
and the programs that use it are handled automatically.

Build Chaos with waf instead of make, adding a tool for sox and for
shared-palette with grit. This also removes the dependency on the
non-Free gfx2gba (which doesn't work on newer Linuxes anyway).

The waf_tools directory is moved from libndspp to the top level and the
configure scripts removed from sub directories.