Add 2 new status for loading pages
This fixes the problem of a image-heavy page being unstoppable. Now we
know when the page is loading or when an internal image is. The progress
bar only shows on the html, not each image (was annoying) and the stop
button no longer flickers stop/refresh/stop/refresh when loading many
images. It sticks on stop until all images are loaded, or stop is pressed.

Also add a stop before previous or next. So if the page is loading, then make
sure we stop() before doing anything else. This prevents problems with the
download queue not finishing.