Add Eclipse CDT project generation
Eclipse project creation that adds all suitable waf targets.

The tool should get the correct local waf and run it using the python
executable that was used to run the version of waf that was used to
generate the project files itself.

Instead of hard coding all the include paths, it takes a waf
BuildContext and extracts the includes from the build objects.